+++ pyogero/const.py
@@ -1,6 +1,7 @@
-""" constants """
-from importlib.resources import files, as_file
-from typing import TypedDict
+"""constants."""
+
+# from importlib.resources import as_file, files
+from typing import NotRequired, TypedDict
 from zoneinfo import ZoneInfo
 
 BASEURL = {
@@ -8,41 +9,38 @@ BASEURL = {
     "myogero": "https://ogero.gov.lb/myogero",
 }
 
-package = 'pyogero.resources'
-resource = 'ogero.pem'
+package = "pyogero.resources"
+resource = "ogero.pem"
 
-with as_file(files(package).joinpath(resource)) as file:
-    CERT_PATH = file
+# with as_file(files(package).joinpath(resource)) as file:
+#     CERT_PATH = file
 
 
 DefaultHeaders = TypedDict(
     "DefaultHeaders",
     {
-        "Accept": str,
-        "Cache-Control": str,
-        "Content-Type": str,
-        "Origin": str,
-        "Referer": str,
+        "Accept": NotRequired[str],
+        "Cache-Control": NotRequired[str],
+        "Content-Type": NotRequired[str],
+        "Origin": NotRequired[str],
+        "Referer": NotRequired[str],
     },
 )
 
 
 def default_headers() -> DefaultHeaders:
-    """returns a default set of headers"""
+    """Return a default set of headers."""
     return {
         "Accept": "application/json",
-        # "Content-Type": "application/json",
-        # "Origin": "https://ogero.gov.lb",
-        # "Referer": "https://ogero.gov.lb/",
         "Cache-Control": "no-cache",
     }
 
 
 API_ENDPOINTS = {
     "login": f'{BASEURL["api"]}/Login.php',
-    "dashboard": f'{BASEURL["myogero"]}/mobileapp.dashboard.php?SessionID={{session_id}}&Username={{username}}&AppRequest&nbr={{phone_account}}&dsl={{internet_account}}',
-    "bill": f'{BASEURL["myogero"]}/bill.php?SessionID={{session_id}}&Username={{username}}&AppRequest&nbr={{phone_account}}',
-    "consumption": f'{BASEURL["myogero"]}/consumption.php?SessionID={{session_id}}&Username={{username}}&AppRequest&dsl={{internet_account}}',
+    "dashboard": f'{BASEURL["myogero"]}/mobileapp.dashboard.php?SessionID={{session_id}}&Username={{username}}&AppRequest&nbr={{phone_account}}&dsl={{internet_account}}',  # noqa: E501
+    "bill": f'{BASEURL["myogero"]}/bill.php?SessionID={{session_id}}&Username={{username}}&AppRequest&nbr={{phone_account}}',  # noqa: E501
+    "consumption": f'{BASEURL["myogero"]}/consumption.php?SessionID={{session_id}}&Username={{username}}&AppRequest&dsl={{internet_account}}',  # noqa: E501
 }
 
 CONNECTION_SPEED = "Current Bundle"
@@ -53,4 +51,7 @@ TOTAL_CONSUMPTION = "Total Consumption"
 EXTRA_CONSUMPTION = "Extra Consumption"
 LAST_UPDATE = "Consumption Until"
 
-LEBANON_TIMEZONE = ZoneInfo('Asia/Beirut')
+LEBANON_TIMEZONE = ZoneInfo("Asia/Beirut")
+
+HTTP_STATUS_OK = 200
+HTTP_STATUS_BAD_REQUEST = 400