+++ tests/test_utils.py
@@ -1,21 +1,21 @@
-""" test utils """
+"""test utils."""
 
 import json
-import os
 from pathlib import Path
 
 import pytest
 
 from pyogero.types import OgeroConfigFile
 
+
 def configloader() -> OgeroConfigFile:
-    """ loads config """
-    for filename in [ os.path.expanduser("~/.config/ogero.json"), "ogero.json" ]:
+    """Load config."""
+    user_path = Path("~/.config/ogero.json")
+    for filename in [Path.expanduser(user_path), "ogero.json"]:
         filepath = Path(filename).resolve()
-        print("filename", filepath)
         if filepath.exists():
             try:
-                with open(filepath, 'r') as f:
+                with Path.open(filepath) as f:
                     return OgeroConfigFile.model_validate_json(f.read())
             except json.JSONDecodeError as json_error:
                 pytest.exit(f"Failed to parse config file: {json_error}")